feat : DAY08 : Mouse Demo 02

+++ Framework/Systems/D3D.cpp
@@ -53,7 +53,7 @@ void D3D::ResizeScreen(float InWidth, float InHeight)
 
 	SAFE_RELEASE(RenderTargetView);
 
-	const HRESULT Hr = SwapChain->ResizeBuffers(0, InWidth, InHeight, DXGI_FORMAT_UNKNOWN, 0);
+	const HRESULT Hr = SwapChain->ResizeBuffers(0, static_cast<UINT>(InWidth), static_cast<UINT>(InHeight), DXGI_FORMAT_UNKNOWN, 0);
 	CHECK(Hr >= 0);
 
 	CreateRTV();
